html/Implantação/autopdf-master/requirements.txt
streamlit==1.28.1
pdfkit==1.0.0
pypdf==3.17.4
PyPDF2==2.12.1
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.108.0
Jinja2==3.1.2
Pillow

